[vfs] E: Split Pipe Domain Types
Move PipeEnd and its private ring-buffer helper into a dedicated module,
and place each public pipe outcome and error enum in its own module.
Preserve the existing vfs::pipe exports and pipe behavior.
